major bugfix: fix wrong output coordinates
Georeferenced position of output raster was shifted to the left and
down, resulting from maths error -- adding and substracting the wrong
variables.

Fixed by using a dictionary for new position values, so that positions
can be referred to by names "xmin, ymax" etc. instead of index numbers
"0,3" which are sooo easy to get mixed up about.