script for ArcGIS export to SVG by Klaus Foerster
"The AsSVG-ArcGIS Geoprocessing script allows to export ArcGIS Point-, Line- or Polygon layers to SVG circle, text or path notation. You may export absolute or relative path d-attributes for Lines and Polygons as well as circles with cx, cy attributes when Relative-Coordinates is checked or text elements with x, y attributes when Relative-Coordinates is unchecked for Points. To reduce File size Coordinate-Precision can be defined. AsSVG does not export styles as set in ArcGIS but allows to specify an attribute via Class-Field that will be used for including classes in the output .svg. All attribute-data specified by Data-Field may be included via custom-namespaced SVG-attributes."

-- http://svg.cc/assvg/arcgis.html, Klaus Foerster <klaus.foerster@svg.cc>